make MixWidgetStateModel public
